+++ BasicFantasyBeyond.Models/CharacterSheetModels/AddCharacterItemsModel.cs
@@ -36,10 +36,10 @@ namespace BasicFantasyBeyond.Models.CharacterSheetModels
         public CharacterClass CharacterClass { get; set; }
         public CharacterAbilities? CharacterAbilities { get; set; }
         public int? CharacterXP { get; set; }
-        public short? CharacterLevel { get; set; }
-        public short? CharacterAC { get; set; }
-        public short? CharacterHP { get; set; }
-        public short? CharacterAttackBonus { get; set; }
+        public int? CharacterLevel { get; set; }
+        public int? CharacterAC { get; set; }
+        public int? CharacterHP { get; set; }
+        public int? CharacterAttackBonus { get; set; }
         public string CharacterNotes { get; set; }
         public IEnumerable<ItemListItem> Items { get; set; }
     }
