+++ BasicFantasyBeyond.Models/CharacterSheetModels/CharacterSheetModel.cs
@@ -42,11 +42,11 @@ namespace BasicFantasyBeyond.Models.CharacterSheetModels
         public CharacterClass CharacterClass { get; set; }
         public CharacterAbilities? CharacterAbilities { get; set; }
         public int CharacterXP { get; set; }
-        public short? CharacterLevel { get; set; }
-        public short? CharacterAC { get; set; }
-        public short? CharacterHP { get; set; }
-        public short? CharacterAttackBonus { get; set; }
-        public short? CharacterSpeed { get; set; }
+        public int? CharacterLevel { get; set; }
+        public int? CharacterAC { get; set; }
+        public int? CharacterHP { get; set; }
+        public int? CharacterAttackBonus { get; set; }
+        public int? CharacterSpeed { get; set; }
         public string CharacterNotes { get; set; }
         public IEnumerable<CharacterItemListItem> Items { get; set; }
         public IEnumerable<CharacterItemListItem> EquippedWeapons { get; set; }
